requirements.txt
streamlit==1.22.0
numpy==1.23.5
Pillow==9.4.0
opencv-python==4.7.0.72
face-recognition==1.3.0
python-dotenv==1.0.0
oauthlib==3.2.2
requests==2.28.2